Host: (0.0.0.0->fqdn); Locking (info->lock)
The server can now be run on a load balancing machine. If this is done,
the host should be read from the info file, not assumed.

+++ zserv/zserv/common.py
@@ -1,7 +1,7 @@
 from functools import wraps
 from . import VERSION as server_version
 from exceptions import VersionMismatchError, AuthenticationRequired
-from settings import INFO_FILE
+from settings import INFO_FILE, LOCK_FILE
 from auth import checkToken, checkTickets, RenewTicketsTimer
 
 def exported(obj):
@@ -35,28 +35,63 @@ def runserver(server):
     signal.signal(signal.SIGHUP, exit)
     signal.signal(signal.SIGTERM, exit)
 
-    info_contents = {"PID": os.getpid(),
-                     "VERSION": server_version}
+    info_contents = {"VERSION": server_version}
     info_contents.update(server.getInfo())
 
+    lck_fd = None
     try:
-        fd = os.open(INFO_FILE, os.O_WRONLY|os.O_CREAT|os.O_EXCL, 0000)
-        fcntl.flock(fd, fcntl.LOCK_EX|fcntl.LOCK_NB)
-        os.fchmod(fd, 0600)
-    except OSError:
-        print "Failed to create info file."
-        exit(1)
-    
-    ticket_timer = RenewTicketsTimer(server.shutdown)
-    try:
-        fobj = os.fdopen(fd, "w")
-        fobj.writelines("%s='%s'\n" % i for i in info_contents.iteritems())
-        fobj.flush()
-        fcntl.flock(fd, fcntl.LOCK_SH|fcntl.LOCK_NB)
-        ticket_timer.start()
-        server.serve_forever()
+        try:
+            lck_fd = os.open(LOCK_FILE, os.O_WRONLY|os.O_CREAT|os.O_EXCL, 0000)
+        except (IOError, OSError):
+            lck_fd = None # Just in case.
+            print "Failed to create lock."
+            exit(1)
+
+        try:
+            fcntl.flock(lck_fd, fcntl.LOCK_EX|fcntl.LOCK_NB)
+        except (IOError, OSError):
+            print "Failed to lock lock."
+            exit(1)
+
+        try:
+            os.write(lck_fd, str(os.getpid()))
+        except (IOError, OSError):
+            print "Failed to write PID to lock."
+            exit(1)
+        
+        try:
+            os.fchmod(lck_fd, 0600)
+        except (IOError, OSError):
+            print "Failed to make lock readable."
+            exit(1)
+        
+        try:
+            with open(INFO_FILE, "w+", 0600) as info_file:
+                info_file.writelines("%s='%s'\n" % i for i in info_contents.iteritems())
+        except IOError:
+            print "Failed to write info file."
+            exit(1)
+
+        try:
+            fcntl.flock(lck_fd, fcntl.LOCK_SH|fcntl.LOCK_NB)
+        except (IOError, OSError):
+            print "Failed to make lock shared."
+            exit(1)
+
+        ticket_timer = RenewTicketsTimer(server.shutdown)
+        try:
+            ticket_timer.start()
+            server.serve_forever()
+        finally:
+            ticket_timer.stop()
     finally:
-        ticket_timer.stop()
-        fobj.close()
-        os.remove(INFO_FILE)
+        if lck_fd is not None:
+            try:
+                os.close(lck_fd)
+            except (OSError, IOError):
+                pass
+        if os.path.exists(INFO_FILE):
+            os.remove(INFO_FILE)
+        if os.path.exists(LOCK_FILE):
+            os.remove(LOCK_FILE)
